requirements.txt
future>=0.16.0
certifi
tornado>=5.1
cryptography
python-telegram-bot
PyGithub